--- src/main/java/domain/NumberPicker.java
@@ -1,5 +1,0 @@
-package domain;
-
-public interface NumberPicker {
-    int pickNumber();
-}
