+++ src/main/java/racingcar/domain/NumberPicker.java
@@ -1,0 +1,5 @@
+package racingcar.domain;
+
+public interface NumberPicker {
+    int pickNumber();
+}
