+++ src/main/java/racingcar/dto/RacingStatusDto.java
@@ -1,0 +1,19 @@
+package racingcar.dto;
+
+public class RacingStatusDto {
+    private final int status;
+    private final String carName;
+
+    public RacingStatusDto(int status, String carName) {
+        this.status = status;
+        this.carName = carName;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public String getCarName() {
+        return carName;
+    }
+}
