+++ src/main/java/racingcar/view/ConsoleOutput.java
@@ -1,0 +1,18 @@
+package racingcar.view;
+
+public class ConsoleOutput implements Output {
+    @Override
+    public void print(String message) {
+        System.out.print(message);
+    }
+
+    @Override
+    public void println(String message) {
+        System.out.println(message);
+    }
+
+    @Override
+    public void println() {
+        System.out.println();
+    }
+}
