+++ src/main/java/racingcar/view/Input.java
@@ -1,0 +1,7 @@
+package racingcar.view;
+
+public interface Input {
+    int gameCountInput();
+
+    String carNameInput();
+}
