+++ src/main/java/racingcar/view/Output.java
@@ -1,0 +1,9 @@
+package racingcar.view;
+
+public interface Output {
+    void print(String message);
+
+    void println(String message);
+
+    void println();
+}
