--- src/main/java/view/Input.java
@@ -1,7 +1,0 @@
-package view;
-
-public interface Input {
-    int gameCountInput();
-
-    String carNameInput();
-}
