--- src/main/java/view/Output.java
@@ -1,9 +1,0 @@
-package view;
-
-public interface Output {
-    void print(String message);
-
-    void println(String message);
-
-    void println();
-}
