+++ src/test/java/racingcar/domain/MockRandomPicker.java
@@ -1,0 +1,11 @@
+package racingcar.domain;
+
+public class MockRandomPicker implements NumberPicker {
+    private int count = 0;
+
+    @Override
+    public int pickNumber() {
+        count = 1 - count;
+        return 3 + count;
+    }
+}
