--- src/test/java/study/CollectionTest.java
@@ -1,47 +1,0 @@
-package study;
-
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.HashSet;
-import java.util.Set;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
-
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-public class CollectionTest {
-
-    private Set<Integer> numbers;
-
-    @BeforeEach
-    void setUp() {
-        numbers = new HashSet<>();
-        numbers.add(1);
-        numbers.add(1);
-        numbers.add(2);
-        numbers.add(3);
-    }
-
-    @Test
-    void size_메서드로_Set_의_크기를_확인한다() {
-        assertThat(numbers.size()).isEqualTo(3);
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {1, 2, 3})
-    void contains_메서드를_이용해_값의_존재를_확인한다(int input) {
-        assertThat(numbers.contains(input)).isTrue();
-    }
-
-    @ParameterizedTest
-    @CsvSource(value = {"1:true", "2:true", "3:true", "4:false", "5:false"}, delimiter = ':')
-    void contains_메서드를_이용해_존재를_부울값으로_확인한다(int input, boolean expected) {
-        assertThat(numbers.contains(input)).isEqualTo(expected);
-    }
-
-}
