--- src/test/java/view/MockInput.java
@@ -1,14 +1,0 @@
-package view;
-
-public class MockInput implements Input {
-
-    @Override
-    public int gameCountInput() {
-        return 3;
-    }
-
-    @Override
-    public String carNameInput() {
-        return "judy,nunu,pobi";
-    }
-}
